refactored subsystem to integrate with MSC service restart/reload
Previously, changes to the CAS configuration profile were entirely
managed by the internal authentication service implementation in
org.soulwing.cas.service.  This made it impractical to create
appropriate service dependencies (in the JBoss MSC) for the deployment
to ensure that a configuration profile could not be removed while in use
by one or more deployments.  The approach in this commit extends base
classes for service reload provided in the Wildfly controller.  A CAS
profile now has a corresponding service in the MSC, and this service is
injected into a per-deployment service when a CAS-enabled web
application is deployed.  In addition to dramatically simplifying the
management model implementation, this makes it easier to manage
additional resource injections such as the SSL context and hostname
verifier.